Added color themes (black, blue and white) which persist between instances, stored locally.
